# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CraftLand Hub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /maps/{mapId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: Get comments for a map
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Post a comment on a map
      parameters:
        - name: mapId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment (admin or owner)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Comment deleted

  # Creators
  /creators:
    get:
      operationId: listCreators
      tags: [creators]
      summary: List all creators
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of creators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Creator"
    post:
      operationId: createCreator
      tags: [creators]
      summary: Create a new creator (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreatorBody"
      responses:
        "201":
          description: Created creator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Creator"
...
components:
  schemas:
...
    Creator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        logo:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
        totalMaps:
          type: integer
        totalLikes:
          type: integer
        isFollowed:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - totalMaps
        - totalLikes
        - isFollowed
        - createdAt
    CreateCreatorBody:
      type: object
      properties:
        name:
          type: string
        logo:
          type: ["string", "null"]
        banner:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        region:
          type: ["string", "null"]
      required:
        - name
...
    Comment:
      type: object
      properties:
        id:
          type: integer
        mapId:
          type: integer
        userId:
          type: string
        username:
          type: ["string", "null"]
        avatar:
          type: ["string", "null"]
        content:
          type: string
        parentId:
          type: ["integer", "null"]
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
        createdAt:
          type: string
      required:
        - id
        - mapId
        - userId
        - content
        - createdAt
        - replies
    CreateCommentBody:
      type: object
      properties:
        content:
          type: string
        parentId:
          type: ["integer", "null"]
      required:
        - content
...